+++ src/main/java/tohru/command/AddEventCommand.java
@@ -1,3 +1,11 @@
+package tohru.command;
+
+import tohru.exception.TohruException;
+import tohru.storage.FileStore;
+import tohru.task.EventItem;
+import tohru.task.TodoList;
+import tohru.ui.Ui;
+
 import java.time.LocalDateTime;
 
 /**
