+++ src/main/java/tohru/command/Command.java
@@ -1,3 +1,10 @@
+package tohru.command;
+
+import tohru.exception.TohruException;
+import tohru.storage.FileStore;
+import tohru.task.TodoList;
+import tohru.ui.Ui;
+
 import java.time.format.DateTimeFormatter;
 
 /**
