+++ src/main/java/tohru/command/ListCommand.java
@@ -1,3 +1,9 @@
+package tohru.command;
+
+import tohru.storage.FileStore;
+import tohru.task.TodoList;
+import tohru.ui.Ui;
+
 /**
  * ListCommand represents the command to list the entries in the to-do list
  */
