+++ src/main/java/tohru/command/Parser.java
@@ -1,3 +1,7 @@
+package tohru.command;
+
+import tohru.exception.TohruException;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
