+++ src/main/java/tohru/task/EventItem.java
@@ -1,3 +1,5 @@
+package tohru.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
