+++ src/main/java/tohru/ui/Ui.java
@@ -1,3 +1,5 @@
+package tohru.ui;
+
 import java.util.Scanner;
 
 /**
